+++ test/e2e/test-proxy-rotation/test.mjs
@@ -38,13 +38,12 @@ await crawler.addRequests(
     ),
 );
 
-const stats = await Actor.main(() => crawler.run(), { exit: false, purge: true });
+const stats = await Actor.main(() => crawler.run(), { exit: false });
 
 expect(stats.requestsFinished === 5, 'All requests finished');
 
 const datasetItems = await getDatasetItems(import.meta.url);
 expect(datasetItems.length >= 5, 'Minimum number of dataset items');
-await delay(1);
 expect(validateDataset(datasetItems, ['clientIp']), 'Dataset items validation');
 
 process.exit(0);